pyproject.toml
[project]
name = "multimodal-agents"
version = "0.1.0"
description = "The 2026 Python Stack for Real-Time Multimodal Agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Stream", email = "support@getstream.io" }]
keywords = ["ai", "multimodal", "agents", "vision", "llm", "openai", "anthropic", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # AI SDKs
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "google-generativeai>=0.5.0",
    "groq>=0.4.0",
    "fireworks-ai>=0.10.0",
    "together>=1.0.0",

    # Input processing
    "opencv-python>=4.8.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "ffmpeg-python>=0.2.0",
    "Pillow>=10.0.0",

    # Async & HTTP
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",

    # Core
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",

    # Benchmarking & visualization
    "matplotlib>=3.7.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
